+++ conftest.py
@@ -16,10 +16,10 @@ from hypothesis import settings
 from pytest import fixture, skip
 
 from lenskit.parallel import ensure_parallel_init
+from lenskit.random import set_global_rng
 
 # bring common fixtures into scope
 from lenskit.testing import ml_100k, ml_ds, ml_ds_unchecked, ml_ratings  # noqa: F401
-from lenskit.util.random import set_global_rng
 
 logging.getLogger("numba").setLevel(logging.INFO)
