+++ docs/api/index.rst
@@ -80,3 +80,4 @@ and may be useful in building new models and components for LensKit.
     lenskit.parallel
     lenskit.testing
     lenskit.util
+    lenskit.random
