+++ lenskit/lenskit/random.py
@@ -219,7 +219,7 @@ def derivable_rng(spec: DerivableSeed) -> RNGFactory:
 
     Seed specifications may be any of the following:
 
-    - A seed (:type:`~lenskit.util.random.SeedLike`).
+    - A seed (:type:`~lenskit.random.SeedLike`).
     - The value ``'user'``, which will derive a seed from the query user ID.
     - A tuple of the form ``(seed, 'user')``, that will use ``seed`` as the
       basis and drive from it a new seed based on the user ID.
