+++ lenskit/lenskit/util/__init__.py
@@ -12,7 +12,7 @@ import logging
 from textwrap import dedent
 from typing import Any, Protocol, TypeVar, runtime_checkable
 
-from .random import derivable_rng, random_generator, set_global_rng
+from ..random import derivable_rng, random_generator, set_global_rng
 from .timing import Stopwatch
 
 try:
